Complete Rework.
The previous version had support for four decks. The bad thing about this
was, that I often needed to page between the controls. But it suits me
more if some controls are accessible at the same time. With a single
Maschine and four full accessible decks, it is not possible in a comfort
way.

This version just supports two decks (and later) the loop recorder.

The Controls:

  Display Area -> EQ Deck A + B
               -> FX 1 + 2

  Groups Area -> Play / Pause (A + B)
              -> Monitor (A + B)
              -> FX 1 + 2 Control (A + B)

  Pad Area -> Set+Activate Loop 2, 4, 8, 16 (A + B)
           -> Loop Move Forward, Back (A + B)
           -> Loop Move Selection (1, 16)

  General -> Browse Collection (Browse button)
          -> Navigate Tree (Volume)
          -> Navigate List (Swing)